refactor: Update login navigation based on user role
This commit updates the login navigation in the Login.jsx file based on the user role retrieved from the API response. If the role is 'admin', the user will be redirected to "/admin". If the role is 'parent', the user will be redirected to "/parent". If the role is 'child', no redirection will occur. This change improves the user experience by providing appropriate navigation based on the user's role.