refactor: Update section height and background color in Events.jsx and home.css
The code changes in Events.jsx and home.css update the section height and background color. In Events.jsx, the section className is modified to include 'section-half-height', which reduces the height of the section. In home.css, the height property of the .sectionBg2 class is changed to 40vh. Additionally, the background color of the navbar in header.css is updated to height: 7rem. These changes improve the visual appearance and layout of the Events and Home pages.