chore: Update package dependencies
This commit updates the package dependencies in the `package.json` and `package-lock.json` files. It adds the `flowbite-react` package with version `0.9.0` and its dependencies. Additionally, it updates the `@floating-ui/react` package to version `0.26.10` and adds the `debounce` package with version `2.0.0`. These updates ensure that the project is using the latest stable versions of the dependencies and includes new features and bug fixes.