chore: Update Toaster component in App.jsx and Users.jsx
This commit updates the Toaster component in the App.jsx and Users.jsx files. In App.jsx, the position prop is set to "bottom-right" and the reverseOrder prop is set to false. In Users.jsx, the Toaster component is commented out to prevent duplicate notifications. These changes improve the positioning and behavior of the toast notifications in the application.