feat: Add helperTrained field to User model and update user routes and UI
This commit adds the `helperTrained` field to the User model in the backend and updates the user routes and UI accordingly. The `helperTrained` field is a boolean value that indicates whether a user has completed helper training. This enhancement allows for better tracking and management of helper training status.

Refactor the user self-update route validations

This commit refactors the user self-update route validations in the backend. The changes include adding new validations for the `helperRegistered` and `disclosureScotland` fields to ensure they are boolean values. This update improves the data integrity and accuracy of the user self-update functionality.

Update admin Users page layout and add helperTrained column

This commit updates the layout of the admin Users page and adds a new column for the `helperTrained` field. This change allows admins to view and manage the helper training status of users more easily.

Update hero section styles for Parent and Events pages

This commit refactors the hero section styles for the Parent and Events pages. The changes include adding an `overlayDash` class to the section and updating the background image and box-shadow properties in the `heroDash` class. These updates enhance the visual appeal and consistency of the pages, providing a better user experience.